docker-python/python/requirements.txt
Flask==1.1.2
psycopg2-binary==2.8.6
gunicorn==20.0.4
